Time: 296 ms (24.36%), Space: 52.9 MB (13.15%) - LeetHub

+++ 0015-3sum/0015-3sum.java
@@ -13,7 +13,7 @@ class Solution {
 				
 			while(s < e)
 			{
-				long sum = (long)(A[i] + A[s] + A[e])	 ;
+				int sum = (A[i] + A[s] + A[e])	 ;
 					
 					if(sum == 0)
 					{
